requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
Django==4.0.3
django-allauth==0.50.0
django-environ==0.8.1
django-phonenumber-field==6.1.0
idna==3.3
oauthlib==3.2.0
phonenumbers==8.12.46
phonenumberslite==8.12.46
Pillow==9.1.0
psycopg==3.0.11
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
urllib3==1.26.9
